+++ addOns/sstiscanner/src/main/java/org/zaproxy/zap/extension/sstiscanner/ReflectedSinkPoint.java
@@ -24,7 +24,7 @@ import org.parosproxy.paros.network.HttpMessage;
 /**
  * Representation of the response to a given request. This can be used to get the state of the sink
  * and store information about it. The main information stored is the response to the original
- * response and the response to a reference request. This information may be useful to compare with
+ * request and the response to a reference request. This information may be useful to compare with
  * the current state of the sink.
  *
  * @author DiogoMRSilva (2018)