+++ addOns/sstiscanner/CHANGELOG.md
@@ -3,6 +3,6 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## First release
 
-- Nothing
+- Add a scanner to find reflected/stored SSTI and one to find  blind SSTI.